requirements.txt
clip==1.0
fschat==0.2.24
pandas==2.0.3
Pillow==9.5.0
Pillow==10.0.0
Pillow==10.0.0
Requests==2.31.0
salesforce_lavis==1.0.1
torchvision==0.15.2
tqdm==4.65.0
